+++ bobtemplates/plone/mockup_pattern/resources/patterns/index.js
@@ -1,0 +1,2 @@
+// Webpack entry point for module federation.
+import "./bundle.js";